@@ -4,10 +4,12 @@
     <Company>Internet Sukces Piotr Stęclik</Company>
     <Copyright>Copyright © Internet Sukces Piotr Stęclik 2016-2024</Copyright>
     <Product>iSukces.Code.Serenity</Product>
-    <TargetFrameworks>netcoreapp3.1;net6.0;net8.0</TargetFrameworks>
+    <ProjectGuid>{2EB071BA-3139-4CA2-93CE-CE7C567A26F9}</ProjectGuid>
+    <TargetFrameworks>net6.0;net8.0;net9.0;netcoreapp3.1</TargetFrameworks>
     <AssemblyVersion>1.23.1218.82</AssemblyVersion>
     <FileVersion>1.23.1218.82</FileVersion>
     <Version>1.23.1218.82</Version>
+    <LangVersion>13</LangVersion>
     <UseWPF>false</UseWPF>
     <SignAssembly>true</SignAssembly>
     <AssemblyOriginatorKeyFile>..\key.snk</AssemblyOriginatorKeyFile>
@@ -17,15 +19,19 @@
     <DocumentationFile>bin\$(Configuration)\$(TargetFramework)\$(AssemblyName).xml</DocumentationFile>
     <GenerateAssemblyDescriptionAttribute>true</GenerateAssemblyDescriptionAttribute>
     <GenerateAssemblyTitleAttribute>true</GenerateAssemblyTitleAttribute>
+    <PackageId>iSukces.Code.Serenity</PackageId>
+    <PackageTags>C#</PackageTags>
+    <Title>Source code generators</Title>
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
+    <ContinuousIntegrationBuild>true</ContinuousIntegrationBuild>
+    <PackageLicenseExpression>MIT</PackageLicenseExpression>
+    <SymbolPackageFormat>snupkg</SymbolPackageFormat>
+    <EmbedUntrackedSources>true</EmbedUntrackedSources>
     <IncludeSource>false</IncludeSource>
     <IncludeSymbols>true</IncludeSymbols>
-    <PackageId>iSukces.Code.Serenity</PackageId>
-    <PackageLicenseExpression>MIT</PackageLicenseExpression>
     <PackageProjectUrl>https://github.com/isukces/iSukces.Code</PackageProjectUrl>
-    <PackageTags>C#</PackageTags>
+    <PublishRepositoryUrl>true</PublishRepositoryUrl>
     <RepositoryUrl>https://github.com/isukces/iSukces.Code</RepositoryUrl>
-    <Title>Source code generators</Title>
   </PropertyGroup>
   <ItemGroup>
     <ProjectReference Include="..\isukces.code\iSukces.Code.csproj" />
